include bridge fee calculation
