+++ governance/helpers/crossChain.js
@@ -5,7 +5,23 @@ const { delayABI } = require('./bridge')
 const { ADDRESS_ZERO, getNetwork } = require('@unlock-protocol/hardhat-helpers')
 const { IConnext } = require('../helpers/bridge')
 
-const parseBridgeCall = async ({ destChainId, moduleData }) => {
+const fetchRelayerFee = async ({ originDomain, destinationDomain }) => {
+  const res = await fetch(
+    'https://sdk-server.mainnet.connext.ninja/estimateRelayerFee',
+    {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({
+        originDomain: originDomain.toString(),
+        destinationDomain: destinationDomain.toString(),
+      }),
+    }
+  )
+  const { hex } = await res.json()
+  return BigInt(hex)
+}
+
+const parseBridgeCall = async ({ srcChainId = 1, destChainId, moduleData }) => {
   const { governanceBridge } = await getNetwork(destChainId)
 
   // get bridge info on receiving chain
@@ -16,12 +32,18 @@ const parseBridgeCall = async ({ destChainId, moduleData }) => {
 
   // get bridge address on mainnet
   const {
-    governanceBridge: { connext: bridgeAddress },
-  } = await getNetwork(1)
+    governanceBridge: { connext: bridgeAddress, domainId: srcDomainId },
+  } = await getNetwork(srcChainId)
 
   if (!destDomainId || !destAddress) {
     throw Error('Missing bridge information')
   }
+
+  // get bridge fee from Connext api
+  const fee = await fetchRelayerFee({
+    destinationDomain: destDomainId,
+    originDomain: srcDomainId,
+  })
 
   // parse call for bridge
   return {
@@ -37,6 +59,8 @@ const parseBridgeCall = async ({ destChainId, moduleData }) => {
       30, // slippage
       moduleData, // calldata
     ],
+    // pay the bridge fee
+    value: fee,
   }
 }
 
@@ -173,6 +197,7 @@ async function simulateDestCalls(xCalls) {
 }
 
 module.exports = {
+  fetchRelayerFee,
   parseBridgeCall,
   simulateDestCalls,
   simulateDelayCall,
